requirements.txt
fastapi
uvicorn
python-magic
PyPDF2
python-dotenv
pdfminer.six
python-multipart